+++ packager/media/codecs/h264_parser.h
@@ -169,13 +169,13 @@ struct H264SliceHeader {
   const uint8_t* nalu_data;
 
   // Size of whole nalu unit.
-  off_t nalu_size;
+  size_t nalu_size;
 
   // This is the size of the slice header not including the nalu header byte.
   // Sturcture: |NALU Header|     Slice Header    |    Slice Data    |
   // Size:      |<- 8bits ->|<- header_bit_size ->|<- Rest of nalu ->|
   // Note that this is not a field in the H.264 spec.
-  off_t header_bit_size;
+  size_t header_bit_size;
 
   int first_mb_in_slice;
   int slice_type;
