+++ packager/media/formats/webm/webm_crypto_helpers.cc
@@ -30,7 +30,7 @@ std::vector<uint8_t> GenerateWebMCounterBlock(const uint8_t* iv, int iv_size) {
 bool WebMCreateDecryptConfig(const uint8_t* data,
                              int data_size,
                              const uint8_t* key_id,
-                             int key_id_size,
+                             size_t key_id_size,
                              std::unique_ptr<DecryptConfig>* decrypt_config,
                              int* data_offset) {
   int header_size = kWebMSignalByteSize;
